mandelbrot in rust
Lol, this is the first time I am actually doing something in rust.

My old mandelbrot (~2yrs ago) was in python and I didn't know proper OOP. Now, I am trying to do this with too much OOP (just look at the complex class). I actually don't know if this is relatively little or a lot of OOP in Rust, but it feels like a lot because I am unfamiliar with it (I am familiar with Java's OOP).